remove ssize_t

+++ src/pyorbbecsdk/types.cpp
@@ -569,7 +569,7 @@ void define_orbbec_types(const py::object &m) {
           "set_data",
           [](OBDataBundle &b, const py::array_t<uint8_t> &arr) {
             if (arr.itemsize() != static_cast<long>(b.itemTypeSize) ||
-                arr.size() != static_cast<ssize_t>(b.itemCount)) {
+                arr.size() != static_cast<long>(b.itemCount)) {
               throw py::value_error("Invalid array size or item size");
             }
             std::memcpy(b.data, arr.data(), arr.nbytes());
